# contract/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger2OpenAPI Converter",
        "version": "1.0.0",
        "contact": {
            "name": "Mike Ralphson",
            "email": "mike.ralphson@gmail.com",
            "url": "https://github.com/mermade/swagger2openapi/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "description": "Converter and validator for Swagger 2.0 to OpenAPI 3.0.x definitions",
        "termsOfService": "https://opensource.org/licenses/MIT"
    },
    "externalDocs": {
        "url": "https://github.com/mermade/openapi-webconverter"
    },
    "schemes": [
        "https",
        "http"
    ],
    "consumes": [
        "application/json",
        "application/x-yaml"
    ],
    "produces": [
        "application/json; charset=utf-8",
        "application/x-yaml"
    ],
    "security": [],
    "host": "openapi-converter.herokuapp.com",
    "basePath": "/api/v1",
    "tags": [
        {
            "name": "metadata",
            "description": "Metadata functions"
        },
        {
            "name": "conversion",
            "description": "Conversion functions"
        },
        {
            "name": "validation",
            "description": "Validation functions"
        }
    ],
    "paths": {
        "/status": {
            "get": {
                "summary": "Get the status of the API",
                "description": "",
                "parameters": [],
                "tags": [
                  "metadata"
                ],
                "operationId": "getStatus",
                "responses": {
                    "200": {
                        "description": "default",
                        "schema": {}
                    }
                }
            }
        },
        "/convert": {
            "get": {
                "summary": "Convert a Swagger 2.0 definition to OpenAPI 3.0.x from a URL",
                "description": "",
                "parameters": [
                    {
                        "name": "url",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "description": "The URL to retrieve the OpenAPI 2.0 definition from",
                        "format": "uri",
                        "x-example": "https://raw.githubusercontent.com/Mermade/openapi-webconverter/master/contract/swagger.json"
                    }
        ],
                "operationId": "convertUrl",
                "tags": [
                    "conversion"
                ],
                "responses": {
                    "200": {
                        "description": "default",
                        "schema": {}
                    }
                }
            },
            "post": {
                "summary": "Convert a Swagger 2.0 definition passed in the body to OpenAPI 3.0.x ",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "filename",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "description": "The file to upload and convert",
                        "x-example": ""
                    },
                    {
                        "name": "source",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "description": "The text of a Swagger 2.0 definition to convert",
                        "x-example": "swagger: 2.0\ninfo:\n  title: API\n  version: 1.0.0\n"
                    },
                    {
                        "name": "validate",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "on"
                        ],
                        "x-example": ""
                    }
                ],
                "operationId": "convert",
                "tags": [
                    "conversion"
                ],
                "responses": {
                    "200": {
                        "description": "default",
                        "schema": {}
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {}
                    }
                }
            }
        },
        "/validate": {
            "get": {
                "summary": "Validate an OpenAPI 3.0.x definition",
                "description": "",
                "parameters": [
                    {
                        "name": "url",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "description": "The URL to retrieve the OpenAPI 3.0.x definition from",
                        "format": "uri",
                        "x-example": "https://raw.githubusercontent.com/Mermade/openapi-webconverter/master/contract/openapi.json"
                    }
                ],
                "operationId": "validateUrl",
                "tags": [
                    "validation"
                ],
                "responses": {
                    "200": {
                        "description": "default",
                        "schema": {}
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "Validate an OpenAPI 3.0.x definition supplied in the body of the request",
                "description": "",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "description": "The file to upload and validate",
                        "x-example": ""
                    },
                    {
                        "name": "source",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "description": "The text of an OpenAPI 3.0.x definition to validate",
                        "x-example": "openapi: 3.0.0\ninfo:\n  title: API\n  version: 1.0.0\npaths: {}\n"
                    }
                ],
                "operationId": "validate",
                "tags": [
                    "validation"
                ],
                "responses": {
                    "200": {
                        "description": "default",
                        "schema": {}
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {}
                    }
                }
            }
        },
        "/badge": {
            "get": {
                "summary": "Return a redirect to a badge svg file depending on a source definition's validity",
                "description": "",
                "parameters": [
                    {
                        "name": "url",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "description": "The URL to retrieve the OpenAPI 3.0.x definition from",
                        "format": "uri",
                        "x-example": "https://raw.githubusercontent.com/Mermade/openapi-webconverter/master/contract/openapi.json"
                    }
                ],
                "operationId": "getBadge",
                "tags": [
                    "validation"
                ],
                "responses": {
                    "301": {
                        "description": "Moved"
                    }
                }
            }
        }
    }
}